Add ESModules import/export for helpers.js
