+++ index.html
@@ -6,7 +6,7 @@
     <title>Calculator</title>
     <link rel="stylesheet" href="style.css">
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
-    <script defer src="script.js"></script>
+    <script type="module" src="script.js"></script>
 </head>
 <body>
    <div id="calculator">
