+++ script.js
@@ -1,13 +1,11 @@
+// const helperFunctions = require("./helpers");
+// console.log("Calculator helper functions loaded:", Object.keys(helperFunctions));
+import {handleCalculatorInput, calculateResult, updateDisplay, clearDisplay, userChoices} from "./helpers.js";
+console.log("Calculator helper functions loaded");
+
 const validOperators = ['+', '-', 'x', '*', '/', '%', '=', 'enter', 'backspace', 'delete', '.', 'shift', 'meta'];
-const display = document.querySelector("#display-text");
 const numPad = document.querySelector("#numpad")
 const calculator = document.querySelector("#calculator");
-
-let userChoices = {
-    firstNumber: '',
-    operation: '',
-    secondNumber: '',
-}
 
 numPad.addEventListener("click", (e) => {
     let target = e.target.textContent;
@@ -53,95 +51,3 @@ calculator.addEventListener("keydown", (e) => {
     updateDisplay();
 
 });
-
-function handleCalculatorInput(target) {
-
-    if (target === '.') {
-        if (isFirstNumber() && (!userChoices.firstNumber.includes('.') && userChoices.firstNumber !== '')) {
-            userChoices.firstNumber += target;
-        } else if (!isFirstNumber() && (!userChoices.secondNumber.includes('.') && userChoices.secondNumber !== '')) {
-            userChoices.secondNumber += target;
-        }
-        return;
-    }
-
-    if (isNumeric(target)) {
-        if (isFirstNumber()) {
-            userChoices.firstNumber += target;
-        } else {
-            userChoices.secondNumber += target;
-        }
-    } else {
-        switch (target) {
-            case '+':
-            case '-':
-            case '/':
-            case '%':
-                if (userChoices.firstNumber !== '') {
-                    userChoices.operation = target;
-                }
-                break;
-            case 'x':
-            case '*':
-                if (userChoices.firstNumber !== '') {
-                    userChoices.operation = 'x';
-                }
-                break; 
-        }
-    }
-}
-
-function calculateResult() {
-    let result;
-    let x = Number(userChoices.firstNumber);
-    let y = Number(userChoices.secondNumber);
-    switch(userChoices.operation) {
-        case '+': 
-            result = x + y;
-            break;
-        case '-': 
-            result = x - y;
-            break;
-        case 'x':
-            result = x * y;
-            break;
-        case '/': 
-            result = x / y;
-            break;
-        case '%': 
-            result = x % y;
-            break;
-    }
-    userChoices.secondNumber = '';
-    userChoices.operation = '';
-    userChoices.firstNumber = String(result);
-};
-
-function updateDisplay() {
-    display.value = `${userChoices.firstNumber}${userChoices.operation}${userChoices.secondNumber}`;
-};
-
-function clearDisplay(target) {
-    switch(target) {
-        case 'C':
-        case 'delete':
-            userChoices.firstNumber = '';
-            userChoices.secondNumber = '';
-            userChoices.operation = '';
-            return;
-        case 'backspace':
-            if (isFirstNumber()){
-                userChoices.firstNumber = userChoices.firstNumber.slice(0,-1);
-            } else if (userChoices.secondNumber === ''){
-                userChoices.operation = userChoices.operation.slice(0,-1);
-            } else {
-                userChoices.secondNumber = userChoices.secondNumber.slice(0,-1);
-            }
-            return;
-        default:
-            return;
-    }
-};
-
-const isNumeric =(value)=> (!isNaN(value) && value !== " ") ? true : false;
-const isFirstNumber = () => (userChoices.operation === '') ? true : false;
